requirements.txt
fastapi
uvicorn
python-multipart
chromadb
httpx
requests
easyocr
pymupdf
python-docx
Pillow
numpy
pydantic
python-dotenv